Add files via upload
Studying the variability of Crime is important as criminology too works on the basis of opportunities. As the changes happen to the lifestyle and routine, the crime scenario too changes accordingly. Therefore, to study this variability of crime requires specific functions that analyse the crime statistics generated. These functions established in this report are one way to analyse and draw interpretations based on geography and the category of the offence. 
The applications of these functions are plenty including the times of emergencies and regular analysis although the limitations of the datasets and the parameters used in the functions are present.